+++ src/mumble/GlobalShortcut_win.h
@@ -34,8 +34,8 @@
 #include "GlobalShortcut.h"
 #include "Timer.h"
 
-#ifdef USE_GKEYS
-#include "GKeys.h"
+#ifdef USE_GKEY
+#include "GKey.h"
 #endif
 
 #define DIRECTINPUT_VERSION 0x0800
@@ -74,7 +74,7 @@ class GlobalShortcutWin : public GlobalShortcutEngine {
 		unsigned int uiHardwareDevices;
 		Timer tDoubleClick;
 		bool bHook;
-#ifdef USE_GKEYS
+#ifdef USE_GKEY
 		GKeyLibrary *gkey;
 #endif
 		static BOOL CALLBACK EnumSuitableDevicesCB(LPCDIDEVICEINSTANCE, LPDIRECTINPUTDEVICE8, DWORD, DWORD, LPVOID);
